requirements.txt
Django==4.2
pyyaml==6.0